cli crate revision
Changed the name to `arbiter`. Changed name of `on-chain` to `onchain`. Completely refactored the `arbiter/src` directory. Added a new directory `./configurations/` that can hold onto config files for us to reduce nesting.
