+++ Cargo.toml
@@ -1,8 +1,8 @@
 [workspace]
 members = [
+    "crates/arbiter",
     "crates/bindings",
-    "crates/cli",
     "crates/simulate",
-    "crates/on-chain",
+    "crates/onchain",
 ]
