+++ crates/arbiter/src/onchain/configuration.rs
@@ -1,0 +1,7 @@
+/// Config object for chian
+#[derive(Serialize, Deserialize, Debug)]
+pub struct ConfigTomlChain {
+    pub contract: String,
+    // RPC url.
+    pub rpc_url: String,
+}
