+++ crates/arbiter/src/simulate/portfolio/arbitrage.rs
