+++ crates/arbiter/src/simulate/portfolio/startup.rs
