+++ crates/arbiter/src/simulate/price_path.rs
