--- crates/cli/src/commands/mod.rs
@@ -1,7 +1,0 @@
-#![warn(missing_docs)]
-#![warn(unsafe_code)]
-//! This module contains the commands for the CLI.
-
-pub mod backtest_data;
-pub mod live;
-pub mod price_path;
