--- crates/cli/src/sim/mod.rs
@@ -1,2 +1,0 @@
-pub mod portfolio;
-pub mod uniswap;
