+++ crates/onchain/src/objectives/linear_nonnegative.rs
