+++ crates/simulate/src/stochastic/price_process.rs
@@ -2,6 +2,7 @@
 
 use plotly::{Plot, Scatter};
 use rand::SeedableRng;
+use serde::{Deserialize, Serialize};
 
 use crate::stochastic::*;
 
@@ -11,8 +12,9 @@ pub trait Plotting {
     fn plot(&self, time: &[f64], price_path: &[f64]);
 }
 
-#[derive(Debug)]
+
 /// Enum for type of price process being used.
+#[derive(Serialize, Deserialize, Debug)]
 pub enum PriceProcessType {
     /// Geometric Brownian Motion (GBM) process.
     GBM(GBM),
@@ -29,6 +31,7 @@ pub enum PriceProcessType {
 /// * `num_steps` - Number of steps in the simulation. (usize)
 /// * `initial_price` - Initial price of the simulation. (f64)
 /// * `seed` - Seed for testing. (u64)
+#[derive(Serialize, Deserialize, Debug)]
 pub struct PriceProcess {
     /// Type of price process.
     pub process_type: PriceProcessType,
@@ -112,7 +115,7 @@ impl Plotting for PriceProcess {
 /// # Fields
 /// * `drift` - Price drift of the underlying asset. (f64)
 /// * `volatility` - Volatility of the underlying asset. (f64)
-#[derive(Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct GBM {
     /// Price drift of the underlying asset.
     drift: f64,
@@ -158,7 +161,7 @@ impl GBM {
 /// # Fields
 /// * `mean_reversion_speed` - Mean reversion speed of the underlying asset. (f64)
 /// * `mean_price` - Mean price of the underlying asset. (f64)
-#[derive(Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct OU {
     /// Volatility of the underlying asset.
     volatility: f64,
